+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -61,9 +61,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-08-19T12:36:38.391Z"
+    created: Optional[str] = "2026-08-20T04:17:08.353Z"
 
-    updated: Optional[str] = "2026-08-19T12:36:38.391Z"
+    updated: Optional[str] = "2026-08-20T04:17:08.353Z"
 
     updated_by_id: OptionalNullable[str] = UNSET