+++ packages/orq-rc/src/orq_ai_sdk/models/patch_v2_human_eval_sets_id_op.py
@@ -228,7 +228,7 @@ class PatchV2HumanEvalSetsIDResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-19T12:36:36.185Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-20T04:17:06.109Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -313,7 +313,7 @@ class PatchV2HumanEvalSetsIDResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-19T12:36:36.185Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-20T04:17:06.109Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")