+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-19T12:36:36.185Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-20T04:17:06.109Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")